+++ server-spi-private/src/main/java/org/keycloak/protocol/trustchain/TrustChainProcessor.java
@@ -2,6 +2,8 @@ package org.keycloak.protocol.trustchain;
 
 import org.keycloak.exceptions.InvalidTrustChainException;
 import org.keycloak.jose.jwk.JSONWebKeySet;
+import org.keycloak.models.IdentityProviderModel;
+import org.keycloak.models.RealmModel;
 import org.keycloak.provider.Provider;
 import org.keycloak.representations.openid_federation.EntityStatement;
 import org.keycloak.representations.openid_federation.TrustChainResolution;
@@ -18,5 +20,7 @@ public interface TrustChainProcessor  extends Provider {
     boolean validateEntityStatementFields(EntityStatement statement, String issuer, String subject);
     void validationRules(EntityStatement statement, boolean checkAudience);
     JSONWebKeySet getKeySet();
+    void updateIdP(IdentityProviderModel model, RealmModel realm);
+    void rPexcplicitRegistration(String opIssuer, String trustAnchor, IdentityProviderModel model, RealmModel realm) throws Exception;
 
 }